+++ CHANGELOG.md
@@ -6,5 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.1] - 2020-11-29
+## [0.0.2] - 2020-11-29
+- make compatible with JIB;
+- configure publishing to the Gradle Plugin Portal;
+- convert key file into property in the plugin;
+- beginning of the documentation
+- cleanup;
+
+## [0.0.1] - 2020-11-29
 - initial working version